- Now all controllers can vibrate given a duration, magnitude, motor, vibration type and intermittence type :)
Added MotorType to specify which motor to vibrate.
Added VibratorMode to specify the kind of vibration and the intermittency to be applied.
Added all necessary vibration fields to the Controller class (all of this needs to be made private!!!)
For efficiency, the vibrations will only update if there is a vibration going on and if we are not paused.
Added the Vibrate, UpdateVibrations, ApplyIntermittence and StopVibration to the Controller class.